+++ include/omw/windows/string.h
@@ -1,6 +1,6 @@
 /*
 author          Oliver Blaser
-date            30.12.2023
+date            31.12.2023
 copyright       MIT - Copyright (c) 2023 Oliver Blaser
 */
 
@@ -30,18 +30,20 @@ namespace omw
 
         std::string wstou8(const wchar_t* src);
         static inline std::string wstou8(const std::wstring& src) { return wstou8(src.c_str()); }
-
-        // TODO deprecated, remove
-        size_t utf8_to_wstr(const char* src, wchar_t* dest, size_t destSize);
-        size_t utf8_to_wstr(const char* src, wchar_t* dest, size_t destSize, omw::windows::ErrorCode& ec);
-        static inline size_t utf8_to_wstr(const std::string& src, wchar_t* dest, size_t destSize) { return utf8_to_wstr(src.c_str(), dest, destSize); }
-        static inline size_t utf8_to_wstr(const std::string& src, wchar_t* dest, size_t destSize, omw::windows::ErrorCode& ec) { return utf8_to_wstr(src.c_str(), dest, destSize, ec); }
-
-        size_t wstr_to_utf8(const wchar_t* src, char* dest, size_t destSize);
-        size_t wstr_to_utf8(const wchar_t* src, char* dest, size_t destSize, omw::windows::ErrorCode& ec);
-        void wstr_to_utf8(const wchar_t* src, std::string& dest);
-        void wstr_to_utf8(const wchar_t* src, std::string& dest, omw::windows::ErrorCode& ec);
         /// @}
+
+        namespace deprecated
+        {
+            size_t utf8_to_wstr(const char* src, wchar_t* dest, size_t destSize);
+            size_t utf8_to_wstr(const char* src, wchar_t* dest, size_t destSize, omw::windows::ErrorCode& ec);
+            static inline size_t utf8_to_wstr(const std::string& src, wchar_t* dest, size_t destSize) { return utf8_to_wstr(src.c_str(), dest, destSize); }
+            static inline size_t utf8_to_wstr(const std::string& src, wchar_t* dest, size_t destSize, omw::windows::ErrorCode& ec) { return utf8_to_wstr(src.c_str(), dest, destSize, ec); }
+
+            size_t wstr_to_utf8(const wchar_t* src, char* dest, size_t destSize);
+            size_t wstr_to_utf8(const wchar_t* src, char* dest, size_t destSize, omw::windows::ErrorCode& ec);
+            void wstr_to_utf8(const wchar_t* src, std::string& dest);
+            void wstr_to_utf8(const wchar_t* src, std::string& dest, omw::windows::ErrorCode& ec);
+        }
 
         /*! @} */
     }